+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   `dismiss_on_scroll` is deprecated
   class names on the popover element have changed
   https://github.com/anvilistas/anvil-extras/pull/563
+* multi-select - rewritten without bootstrap, now uses popover and basic anvil components
+  https://github.com/anvilistas/anvil-extras/pull/566
 
 
 # v2.7.0 05-Sep-2024